AArch64: Improve C1 Stub Frame code
For many of the stub frames, a leave/ret is generated after the stub has
already branched or returned. This is confusing. For these cases, replace
the superfluous code with a should_not_reach_here

For handle excception, instead of storing return from the exception
handler on the stack, it can be moved directly into lr, replacing a store and
load with a single move. (If/when PAC support is implemented, then this store
would also have to be signed).