Allow AsyncGetCallTrace and JFR to walk a stub frame
When the signal sent for AsyncGetCallTrace or JFR would land on a stub
(like arraycopy), it wouldn't be able to detect the sender (caller)
frame because `_cb->frame_size() == 0`.

Because we fully control how the prolog and epilog of stub code is
generated, we know there are two cases:
1. A stack frame is allocated via macroAssembler->enter(), and consists
in `push rbp; mov rsp, rbp;`.
2. No stack frames are allocated and rbp is left unchanged and rsp is
decremented with the `call` instruction that push the return `pc` on the
stack.

For case 1., we can easily know the sender frame by simply looking at
rbp, especially since we know that all stubs preserve the frame pointer
(on x86 at least).

For case 2., we end up returning the sender's sender, but that already
gives us more information than what we have today.

The results are as follows:

- Baseline:

Flat Profile (by method):
        (t 99.4,s 99.4) AGCT::Unknown Java[ERR=-5]
        (t  0.5,s  0.2) Prof1::main
        (t  0.2,s  0.2) java.lang.AbstractStringBuilder::append
        (t  0.1,s  0.1) AGCT::Unknown not Java[ERR=-3]
        (t  0.0,s  0.0) java.lang.AbstractStringBuilder::ensureCapacityInternal
        (t  0.0,s  0.0) java.lang.AbstractStringBuilder::shift
        (t  0.0,s  0.0) java.lang.String::getBytes
        (t  0.0,s  0.0) java.lang.AbstractStringBuilder::putStringAt
        (t  0.0,s  0.0) java.lang.StringBuilder::delete
        (t  0.2,s  0.0) java.lang.StringBuilder::append
        (t  0.0,s  0.0) java.lang.AbstractStringBuilder::delete
        (t  0.0,s  0.0) java.lang.AbstractStringBuilder::putStringAt

- With StubRoutinesBlob::FrameParser

Flat Profile (by method):
        (t 98.7,s 98.7) java.lang.AbstractStringBuilder::ensureCapacityInternal
        (t  0.9,s  0.9) java.lang.AbstractStringBuilder::delete
        (t 99.8,s  0.2) Prof1::main
        (t  0.1,s  0.1) AGCT::Unknown not Java[ERR=-3]
        (t  0.0,s  0.0) AGCT::Unknown Java[ERR=-5]
        (t 98.8,s  0.0) java.lang.AbstractStringBuilder::append
        (t 98.8,s  0.0) java.lang.StringBuilder::append
        (t  0.9,s  0.0) java.lang.StringBuilder::delete

The program is as follows:

```
public class Prof1 {

    public static void main(String[] args) {
        StringBuilder sb = new StringBuilder();
        for (int i = 0; i < 1000000; i++) {
            sb.append("ab");
            sb.delete(0, 1);
        }
        System.out.println(sb.length());
    }
}
```

We now account for the arraycopy stub which is called by
AbstractStringBuilder::ensureCapacityInternal. It was previously ignored
because it would not know how to parse the frame for the arraycopy stub
and would fall in the AGCT::Unknown Java[ERR=-5] section.

However, it still isn't perfect since it doesn't point to the arraycopy stub
directly.