8268178: Extract sender frame parsing to CodeBlock::FrameParser
Whether and how a frame is setup is controlled by the code generator
for the specific CodeBlock. The CodeBlock is then in the best place to know how
to parse the sender's frame from the current frame in the given CodeBlock.

This refactoring proposes to extract this parsing out of `frame` and into a
`CodeBlock::FrameParser`. This FrameParser is then specialized in the relevant
inherited children of CodeBlock.

This change is to largely facilitate adding new supported cases for JDK-8252417
like runtime stubs.