stubGenerator_ppc.cpp: decodeBlock(): Use xxpermx to improve performance of decodeBlock on Power10+
xxpermx, available in Power64 ISA 3.1+, provides the ability to do
lower-cost, large table look-ups.  Here we use it for looking up a 7-bit
Base64 character and translating it to a 6-bit value, while also detecting
illegal Base64 characters.  Because Base64 characters are restricted to
values between 32 and 127, we only need three xxpermx instructions to
complete the lookup.

This yields about a 5% improvement in performance, with a small amount
additional overhead per intrinsic call due to need to load constants into
the lookup table vector registers.