JDK-8269092: Add OldObjectSampleEvent.allocationSize field
The data is already available in ObjectSample.allocated and getting this data
in another way is complicated (via Instrumentation for example). Having the
size of the objects also allows to augment the memory leak analysis based on
the object size impact on top of based on the object count impact.