Add xxpermx instruction, and align_prefix() method
In addition to the definition of the xxpermx instruction, this commit also
adds align_prefix() which conditionally adds a nop if needed to push an
instruction prefix past the 64-byte cache line boundary as required by the
ISA 3.1 architecture.  Previous commits to add prefixed instructions didn't
support their use from intrinsics, because instruct nodes in ppc.ad can
specify their own alignment requirements.  Since we can't make use of
instruct mechanism in intrinsics, I have added align_prefix(), which needs
to be called explicitly for each use of a prefixed instruction.  I feel
this is a controversial thing to do, because it requires that anyone using
a prefixed instruction in an intrinsic will need to remember to add a call
to align_prefix() before each instruction.  We had a discussion here about
this, but haven't come up with any alternatives that seem to be workable.