stubGenerator_cpp.cpp: decodeBlock(): use constant block for loading constants into vector registers for cleaner and faster code
The code is cleaner because we only need to load one address, and then all
constants are loaded from that address plus a specific offset.

This reduces the overhead of calling the instrinsic, because although the
constant loading isn't in the main loop, it is performed on every call to
the intrinsic.