8268363: AArch64: Implement string_indexof_char intrinsic in SVE
This patch implements string_indexof_char intrinsic in SVE for both
UTF-16 and Latin-8 variants. In the current underneath implementation,
there is only 4 and 8 characters are compared at a time in UTF-16 and
Latin-8 respectively. In this patch, we adopt SVE to vectorize this
intrinsic so that a wider comparison can be performed each time.

I've run [1] with this patch on an SVE machine, and the performance of
StringIndexOfChar.latin1_mixed_char (length=65) and
StringIndexOfChar.utf16_mixed_char (length=65) are improved significantly,
while no visible regression for other cases with shorter strings.
(length=7, 8, 15, 16 and 32)

[1] https://github.com/openjdk/jdk/blob/master/test/micro/org/openjdk/bench/java/lang/StringIndexOfChar.java