Increase stride size to allow unrolling more loops
This increase allowed stride size for loop unrolling to almost maxium possible
value which is around `max_jint /  2 - 2`,
so the value which will prevent overflow when stride is doubled in C2.

The motivation of this change is discussion and research about unrolling
vector (SIMD) loops. In such a case stride size depends on elements size
and machine size of vector for AVX256 and int stride size is 8,
and loop unroll happens. However short vectors will not cause loop unroll.

```
    for (int i = 0; i < SPECIES.loopBound(longSize); i += SPECIES.length() /* 8 for int, 16 for short */ ) {
      var v = ShortVector.fromByteBuffer(SPECIES, srcBufferHeap, i << 1, ByteOrder.nativeOrder());
      v.intoByteBuffer(dstBufferHeap, i << 1, ByteOrder.nativeOrder());
    }
```