8271323: [TESTBUG] serviceability/sa/ClhsdbCDSCore.java fails with -XX:TieredStopAtLevel=1
This test fails reliably with -XX:TieredStopAtLevel=1 since JDK-8251462.
MDOs are no longer allocated in this mode so the clhsdb printmdo -a
command prints nothing.  The failure is basically the same as
JDK-8236042 which happened with -Xcomp -XX:TieredStopAtLevel=1, so the
workaround for that just needs to be adjusted slightly.