Add @summary to tests
From https://openjdk.java.net/jtreg/faq.html:

The @summary tag describes the condition that is checked by the test. It is especially useful for non-regression tests, which by definition don't have bug numbers, but even if there's a bug number it's helpful to include a summary. Note that a test summary is generally not the same thing as a Bugtraq synopsis, since the latter describes the bug rather than the condition that the bug violates.

That said, the JBS synopsis for 8266666 suits those tests just fine.