Add concept of Filtering based on RecordingContext
This prototype allows to filter on the content of a context, without
having to recompute the filter on every event but only accessing a TLS
variable.

However, it doesn't allow to filter on the presence or not of said
context. That would need to store whether the filters are matched not
on the RecordingContextBinding, but somewhere more agnostic.