8276985: AArch64: [vectorapi] Backend support of VectorMaskToLongNode
The lack of codegen for VectorMaskToLong results in a regression on
AArch64 for VectorMask.laneIsSet, which relies on the intrinsification
of VectorMask.toLong after JDK-8273949.

This patch implements bitmask extraction on AArch64 for NEON and SVE by
using scalar instructions, which is equivalent to the PMOVMSK
instructions on X86. The performance of VectorMask.laneIsSet improves
about 10x for NEON and 2x~4x for SVE on my test machines.

Change-Id: I1be2e107c19344a095afcb9e807f468b47ae2dfe