8277619: AArch64: Incorrect parameter type in Advanced SIMD Copy assembler functions
Mov (from general), incorrectly uses SIMD_Arrangement as the parameter
type of the assembler function. However, from Arm ARM [1], it's more
precise to use SIMD_RegVariant here.

The situation is similar to Mov(to general) [2].

Note that as Mov(to general) is an alias of UMOV, we turn to re-use
UMOV encoding for Mov(to general) in this patch.

[1] https://developer.arm.com/documentation/ddi0602/2020-12/SIMD-FP-Instructions/MOV--from-general---Move-general-purpose-register-to-a-vector-element--an-alias-of-INS--general--
[2] https://developer.arm.com/documentation/ddi0602/2020-12/SIMD-FP-Instructions/MOV--to-general---Move-vector-element-to-general-purpose-register--an-alias-of-UMOV-

Change-Id: Ieafd0ce0f82814b606701863a6a47d345d5a6ffd