8176501: Method Shape.getBounds2D() incorrectly includes Bezier control points in bounding box
Make the new getBounds2D(PathIterator) method package private.

This is in response to prrace's code review feedback:

> public static Rectangle2D getBounds2D(final PathIterator pi);
>
> Is this really necessary ? It is just for the benefit of the
> public API caller so can be package private.

https://github.com/openjdk/jdk/pull/6227#issuecomment-995305817

Personally (having written custom java.awt.Shape classes) I think this method would be of general use (so it could be public), but if that's the minority opinion: I'm happy to convert it to package private for now. If we ever want to discuss making it public then that could be a separate PR/conversation.