Fix and simplify executableMembersEqual
Instances of VariableElement should be compared using their types, not their type FQNs. For a type variable v, its FQN ( getTypeName(v.asType(), true) ) seems to degrade to the variable name. For example, when comparing static Set.of(E) to static EnumSet.of(E), executableMembersEqual used to return true because it didn't consider the type of Es. Although variables are called the same (E), for Set.of E is anything, whereas for EnumSet.of E is an enum.