Use appropriate warnings
Using `inheritDoc` on a tag that does not support inheritance differs from using `inheritDoc` on a method that is neither overridden nor implemented. Broadly speaking, there are two cases of that tag misuse:

  1. using inheritDoc where it can never be used
  2. using inheritDoc where it can generally be used but not in this particular case

We could live with just the more general message, but then it would be less helpful in more specific cases.