8285630: Fix a configure error in RISC-V cross build
We are trying to cross build a RISC-V version of OpenJDK. We specified
`--openjdk-target=riscv64-linux-gnu` after `bash configure` but got an
error message.

```
configure: error: /usr/bin/bash /home/ent-user/jdk_src/make/autoconf/build-aux/config.sub riscv64-linux-gnu failed
configure exiting with result code 1
```

It shows the processing of `riscv64-linux-gnu` in script `config.sub`
fails. We see `config.sub` calls another script `autoconf-config.sub`
to validate and canonicalize a configuration type. The validation fails
here because `autoconf-config.sub` is a quite old version and has no
RISC-V support inside. Comments in those scripts tell us patching the
autoconf script is not a good idea so we add a fix in `config.sub` in
this patch.

We have verified RISC-V cross build succeeds after this change. As we
are not quite familiar with the build system, please let us know if this
is the best way to fix.