8286711: AArch64: serviceability agent tests fail with PAC enabled
When the VM is run with `-XX:UseBranchProtection=pac-ret` on a supported
CPU, the upper bits of the saved link register contain a "pointer
authentication code" which must be checked and removed by a special
instruction before a function returns.  The serviceability agent is
unaware of this and so sees garbled PC values when walking the stack.
This patch just masks out the PAC bits from the sender PC after reading
it from the stack frame.