8287325: AArch64: fix virtual threads with -XX:UseBranchProtection=pac-ret
The continuation free/thaw mechanism relies on being able to move thread
stacks around in memory.  However when PAC is enabled on supported
AArch64 CPUs, the saved LR on the stack contains a "pointer
authentication code" signed with the stack pointer at the time the frame
was created.  When a stack frame is relocated we need to re-sign the LR
with the new stack pointer to ensure it authenticates successfully when
the method returns.

Also introduced ContinuationHelper::return_pc_at() to avoid directly
reading the saved PC from the stack in shared code.  On AArch64 with PAC
it enabled it strips the PAC from the address after reading it, on all
other platforms it just loads the PC from the stack as before.