8176501: Method Shape.getBounds2D() incorrectly includes Bezier control points in bounding box
Modify the unit test so we log random seeds. Because these unit tests use random numbers: failures can appear randomly. With this commit: we'll log the random seeds so we can easily reproduce specific failures. Also we'll be able to say with certainty if we fixed a failure or not.