8176501: Method Shape.getBounds2D() incorrectly includes Bezier control points in bounding box
This resolves occasional unit test failures in java.awt.geom.Path2D.UnitTest. (For ex, see seeds 4603421469924484958L, 4596019360892069260L, 4604586530476373958L, 4603766396818608126L)

The problem appears to be that the UnitTest class randomly creates PathIterator segments in a way that can place lines/quads/cubics immediately after a SEG_CLOSE. That is: after a SEG_CLOSE there was no SEG_MOVETO as new segments followed.

The previous Path2D#getBounds2D(PathIterator) method assumed each subpath *always* started with a SEG_MOVETO. (I don't know if this is formally stated in a specification anywhere or not?) The unit test failures show this is a bad assumption.

With this change: when a SEG_CLOSE is encountered: we reset the cursor to the last (moveX, moveY) point. This appears to resolve the unit test failures, which suggests this is consistent with what existing awt.geom classes also do.