TIFF plugin: JPEG compression and alpha channel is not supported
This is a followup on email thread:
https://mail.openjdk.java.net/pipermail/client-libs-dev/2022-May/004882.html

The gist is that having transparency (alpha) in a JPEG compressed TIFF
container is not supported. Unsupported Image Type exception used to be
thrown before JDK-8274735 [openjdk/jdk/pull/7849] got integrated.
Since then, the 4 components JPEG from such TIFF file is erroneously
interpreted as CMYK, while in fact it is an unsupported RGBA.

This commit makes the TIFFImageReader to throw a meaningful exception
informing user that such combination of colour space and compression
is not supported.

The accompanying test verifies that the exception is indeed thrown
when expected.