JDK-8264999: GeneralPath.lineTo() to itself produces jagged lines
Adding "&& prev == DRAWING_OP_TO" per Laurent's code review.

I also redid the unit tests. Before this change, with the new unit tests: one scenario failed. With this change: four scenarios failed.

For now I've commented out the failing tests. The failing tests are not the same as the original bug; they are related considerations we should also test against, but IMO they are not essential for the resolution of jdk-8264999.